Resume box and search box integrated with the new api data.
