+++ src/newPages/statistics/students/Actives/activeSlider.css
