+++ src/newPages/statistics/students/Actives/styles.css
@@ -125,6 +125,12 @@
 .resume-box>div:nth-child(4){
     border-bottom: 8px solid #FF0000;
 }
+.resume-box>div:nth-child(5){
+    border-bottom: 8px solid #85847d;
+}
+.resume-box>div:nth-child(6){
+    border-bottom: 8px solid #000000;
+}
 
 /* from the second child */
 .resume-box div:nth-child(n+2){
@@ -158,7 +164,7 @@
     box-sizing: border-box;
 
     /* width: 70px; */
-    font-size: 1rem;
+    font-size: .8rem;
     font-weight: 500;
 }
