--- src/pages/newDesign/Actives/TableDiscent/styles.css
@@ -1,4 +1,0 @@
-.capitalize, th{
-    text-transform: capitalize;
-}
-
